fix: BCK-044 Added comments folder for serializers and views and tests
